+++ app/src/main/java/org/linphone/activities/main/contact/data/ContactEditorData.kt
@@ -17,43 +17,32 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.linphone.activities.main.contact.viewmodels
+package org.linphone.activities.main.contact.data
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.media.ExifInterface
 import android.provider.ContactsContract
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewModelScope
 import java.io.ByteArrayOutputStream
 import java.io.IOException
 import kotlinx.coroutines.CoroutineScope
 import org.linphone.LinphoneApplication.Companion.coreContext
 import org.linphone.LinphoneApplication.Companion.corePreferences
-import org.linphone.activities.main.contact.data.NumberOrAddressEditorData
+import org.linphone.R
 import org.linphone.contact.*
 import org.linphone.core.ChatRoomSecurityLevel
 import org.linphone.core.Friend
 import org.linphone.core.tools.Log
+import org.linphone.utils.AppUtils
 import org.linphone.utils.ImageUtils
 import org.linphone.utils.PermissionHelper
 
-class ContactEditorViewModelFactory(private val friend: Friend?) :
-    ViewModelProvider.NewInstanceFactory() {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return ContactEditorViewModel(friend) as T
-    }
-}
-
-class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface {
+class ContactEditorData(val friend: Friend?) : ContactDataInterface {
     override val contact: MutableLiveData<Friend> = MutableLiveData<Friend>()
     override val displayName: MutableLiveData<String> = MutableLiveData<String>()
     override val securityLevel: MutableLiveData<ChatRoomSecurityLevel> = MutableLiveData<ChatRoomSecurityLevel>()
-    override val coroutineScope: CoroutineScope = viewModelScope
+    override val coroutineScope: CoroutineScope = coreContext.coroutineScope
 
     val firstName = MutableLiveData<String>()
 
@@ -74,20 +63,20 @@ class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface
     var syncAccountType: String? = null
 
     init {
-        if (c != null) {
-            contact.value = c!!
-            displayName.value = c.name ?: ""
+        if (friend != null) {
+            contact.value = friend!!
+            displayName.value = friend.name ?: ""
         } else {
             displayName.value = ""
         }
 
-        organization.value = c?.organization ?: ""
+        organization.value = friend?.organization ?: ""
 
         firstName.value = ""
         lastName.value = ""
-        val vCard = c?.vcard
+        val refKey = friend?.refKey
+        val vCard = friend?.vcard
         if (vCard?.familyName.isNullOrEmpty() && vCard?.givenName.isNullOrEmpty()) {
-            val refKey = c?.refKey
             if (refKey != null) {
                 Log.w("[Contact Editor] vCard first & last name not filled-in yet, doing it now")
                 fetchFirstAndLastNames(refKey)
@@ -99,11 +88,11 @@ class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface
             lastName.value = vCard?.familyName
         }
 
-        updateNumbersAndAddresses()
+        updateNumbersAndAddresses(refKey)
     }
 
     fun save(): Friend {
-        var contact = c
+        var contact = friend
         var created = false
 
         if (contact == null) {
@@ -144,7 +133,7 @@ class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface
         }
         for (address in addresses.value.orEmpty()) {
             val sipAddress = address.newValue.value.orEmpty()
-            if (sipAddress.isEmpty()) continue
+            if (sipAddress.isEmpty() || address.toRemove.value == true) continue
 
             val parsed = coreContext.core.interpretUrl(sipAddress, false)
             if (parsed != null) contact.addAddress(parsed)
@@ -155,7 +144,7 @@ class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface
         }
         for (phone in numbers.value.orEmpty()) {
             val phoneNumber = phone.newValue.value.orEmpty()
-            if (phoneNumber.isEmpty()) continue
+            if (phoneNumber.isEmpty() || phone.toRemove.value == true) continue
 
             contact.addPhoneNumber(phoneNumber)
         }
@@ -228,20 +217,71 @@ class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface
         numbers.value = list
     }
 
-    private fun updateNumbersAndAddresses() {
+    private fun updateNumbersAndAddresses(contactId: String?) {
         val phoneNumbers = arrayListOf<NumberOrAddressEditorData>()
-        for (number in c?.phoneNumbers.orEmpty()) {
-            phoneNumbers.add(NumberOrAddressEditorData(number, false))
+        val sipAddresses = arrayListOf<NumberOrAddressEditorData>()
+        var fetched = false
+
+        if (contactId != null) {
+            try {
+                // Try to get real values from contact to ensure edition/removal in native address book will go well
+                val cursor = coreContext.context.contentResolver.query(
+                    ContactsContract.Data.CONTENT_URI,
+                    arrayOf(
+                        ContactsContract.Data.MIMETYPE,
+                        ContactsContract.CommonDataKinds.Phone.NUMBER
+                    ),
+                    ContactsContract.CommonDataKinds.StructuredName.CONTACT_ID + " = ?",
+                    arrayOf(contactId),
+                    null
+                )
+
+                while (cursor != null && cursor.moveToNext()) {
+                    val linphoneMime = AppUtils.getString(R.string.linphone_address_mime_type)
+                    val mime: String? =
+                        cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.Data.MIMETYPE))
+                    if (mime == ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE) {
+                        val data1: String? =
+                            cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.NUMBER))
+                        if (data1 != null) {
+                            phoneNumbers.add(NumberOrAddressEditorData(data1, false))
+                        }
+                    } else if (
+                        mime == ContactsContract.CommonDataKinds.SipAddress.CONTENT_ITEM_TYPE ||
+                        mime == linphoneMime
+                    ) {
+                        val data1: String? =
+                            cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.SipAddress.SIP_ADDRESS))
+                        if (data1 != null) {
+                            sipAddresses.add(NumberOrAddressEditorData(data1, true))
+                        }
+                    }
+                }
+
+                cursor?.close()
+                fetched = true
+            } catch (e: Exception) {
+                Log.e("[Contact Editor] Failed to sip addresses & phone number: $e")
+                fetched = false
+            }
         }
+
+        if (!fetched) {
+            Log.w("[Contact Editor] Fall-backing to friend info (might be inaccurate and thus edition/removal might fail)")
+            for (number in friend?.phoneNumbers.orEmpty()) {
+                phoneNumbers.add(NumberOrAddressEditorData(number, false))
+            }
+
+            for (address in friend?.addresses.orEmpty()) {
+                sipAddresses.add(NumberOrAddressEditorData(address.asStringUriOnly(), true))
+            }
+        }
+
         if (phoneNumbers.isEmpty()) {
             phoneNumbers.add(NumberOrAddressEditorData("", false))
         }
         numbers.value = phoneNumbers
 
-        val sipAddresses = arrayListOf<NumberOrAddressEditorData>()
-        for (address in c?.addresses.orEmpty()) {
-            sipAddresses.add(NumberOrAddressEditorData(address.asStringUriOnly(), true))
-        }
         if (sipAddresses.isEmpty()) {
             sipAddresses.add(NumberOrAddressEditorData("", true))
         }
@@ -268,14 +308,14 @@ class ContactEditorViewModel(val c: Friend?) : ViewModel(), ContactDataInterface
                     val givenName: String? =
                         cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.StructuredName.GIVEN_NAME))
                     if (!givenName.isNullOrEmpty()) {
-                        c?.vcard?.givenName = givenName
+                        friend?.vcard?.givenName = givenName
                         firstName.value = givenName!!
                     }
 
                     val familyName: String? =
                         cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.StructuredName.FAMILY_NAME))
                     if (!familyName.isNullOrEmpty()) {
-                        c?.vcard?.familyName = familyName
+                        friend?.vcard?.familyName = familyName
                         lastName.value = familyName!!
                     }
                 }